backend/requirements.txt
# FastAPI Backend Requirements for Ocean Data Integration Platform
# Smart India Hackathon 2024

# FastAPI and ASGI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0

# Supabase
supabase>=2.0.0
postgrest>=0.13.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Data Processing
pandas>=1.5.0
numpy>=1.21.0
pydantic>=2.0.0
pydantic[email]>=2.0.0

# Machine Learning
scikit-learn>=1.1.0
joblib>=1.2.0

# HTTP Client
httpx>=0.25.0
requests>=2.28.0

# Utilities
python-dateutil>=2.8.0
pytz>=2022.1

# Development & Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0

# Optional: Redis for caching
redis>=4.5.0
